+++ tests/testthat/test-add_phylopic_base.R
@@ -4,15 +4,16 @@ test_that("add_phylopic_base works", {
   # phylopic in background, with name
   expect_doppelganger("phylopic in background", function() {
     plot(1, 1, type = "n", main = "A cat")
-    add_phylopic_base(name = "Cat")
+    add_phylopic_base(name = "Felis silvestris")
   })
 
   # png phylopic in background
   expect_doppelganger("phylopic png in background", function() {
     cat_png <- get_phylopic("23cd6aa4-9587-4a2e-8e26-de42885004c9",
-                            format = "512")
+                            format = "raster")
     plot(1, 1, type = "n", main = "A cat")
-    add_phylopic_base(cat_png, x = 1, y = 1, ysize = .4, color = "blue")
+    add_phylopic_base(cat_png, x = 1, y = 1, ysize = .4, color = "blue",
+                      alpha = .5, angle = -90, horizontal = TRUE)
   })
 
   # a bunch of little colored phylopics, with existing image
@@ -20,16 +21,20 @@ test_that("add_phylopic_base works", {
   posx <- runif(50, 0, 10)
   posy <- runif(50, 0, 10)
   sizey <- runif(50, 0.4, 2)
+  angle <- runif(50, 0, 360)
+  hor <- sample(c(TRUE, FALSE), 50, TRUE)
+  ver <- sample(c(TRUE, FALSE), 50, TRUE)
   cols <- sample(c("black", "darkorange", "grey42", "white"), 50,
                  replace = TRUE)
+  alpha <- runif(50, 0, 1)
 
-  cat <- get_phylopic("23cd6aa4-9587-4a2e-8e26-de42885004c9")
   expect_doppelganger("phylopics on top of plot", function() {
     plot(posx, posy, type = "n", main = "A cat herd")
-    for (i in 1:50) {
-      add_phylopic_base(cat, x = posx[i], y = posy[i], ysize = sizey[i],
-                        color = cols[i])
-    }
+    add_phylopic_base(uuid = "23cd6aa4-9587-4a2e-8e26-de42885004c9",
+                      x = posx, y = posy, ysize = sizey,
+                      color = cols, alpha = alpha,
+                      angle = angle,
+                      horizontal = hor, vertical = ver)
   })
 
   # Expect error
@@ -38,6 +43,5 @@ test_that("add_phylopic_base works", {
   expect_error(add_phylopic_base())
   expect_error(add_phylopic_base(cat, alpha = 3))
   expect_error(add_phylopic_base(name = 42))
-  expect_error(add_phylopic_base(name = "bueller"))
   expect_error(add_phylopic_base(uuid = 42))
 })