+++ NEWS.md
@@ -1,3 +1,19 @@
+rphylopic 1.1.0
+===============
+
+* added functions for transforming PhyloPic silhouettes (flipping and rotating)
+* save_phylopic bg argument updated to be "transparent" by default
+* added geom_phylopic (#25)
+* vectorized add_phylopic and add_phylopic_base (#42)
+* recolor_phylopic now removes white backgrounds by default
+* fixed the handling of alpha values
+* get_phylopic can now return any size raster image (#50)
+* removed the "thumbnail" and "twitter" format options for get_phylopic
+* fixed how silhouettes are gathered from PhyloPic (#51)
+* pick_phylopic updated to allow visualization of multiple silhouettes at once (#43)
+* fixed add_phylopic_base for multi-panel figures
+* added three vignettes (#49, #55)
+
 rphylopic 1.0.0
 ==============